Add mode parameter to plaits
The return value of MiPlaits is an array of two signals. Unlike most
multiple-output UGens, these don't represent stereo signals; instead, the first
signal is designated the primary output and the second, the auxiliary output (or
AUX in the docs) represents a skewed/altered/lofi version of the primary output.
Before this patch, these signals were getting mixed together, which is why some
of these drums sounded so weird. Introducing the `mode` parameter allows us
to toggle between the primary and auxiliary channels.

This is a backwards-incompatible change to the synth, but you can recover the
original behavior by passing `[0|1]` as the `mode` parameter.

Additionally, we can remove the `SplayAz` trick, as the presence of the `Select`
UGen cuts `sig` down to one signal, and `DirtPan` takes care of splaying a
single channel across the available channels.